+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/Shared/PrometheusExporter.cs
@@ -28,6 +28,7 @@ internal sealed class PrometheusExporter : BaseExporter<Metric>, IPullMetricExpo
         this.TargetInfoEnabled = options.TargetInfoEnabled;
         this.DisableTotalNameSuffixForCounters = options.DisableTotalNameSuffixForCounters;
         this.ResourceConstantLabels = options.ResourceConstantLabels;
+        this.MaxScrapeResponseSizeBytes = options.MaxScrapeResponseSizeBytes;
 
         this.CollectionManager = new PrometheusCollectionManager(this);
     }
@@ -54,6 +55,8 @@ internal sealed class PrometheusExporter : BaseExporter<Metric>, IPullMetricExpo
     internal bool DisableTotalNameSuffixForCounters { get; }
 
     internal Func<string, bool>? ResourceConstantLabels { get; }
+
+    internal int MaxScrapeResponseSizeBytes { get; }
 
     internal Resource Resource
     {